refactor: extract constants to const.ts and utilities to utils.ts
Move annotation/label constants and ConfigType enum from types.ts to a
dedicated const.ts. Move isConfigEnabled, isConfigPreInstalled and their
helper hasKserveOwnership from types.ts to utils.ts. Re-export all
constants from types.ts to preserve backwards compatibility.

Addresses review feedback that utility functions don't belong in a types
file.
