+++ packages/llmd-serving/src/const.ts
@@ -1,0 +1,10 @@
+// --- Label/Annotation Constants (dashboard-owned, abstracted for easy refactor) ---
+
+export const MAAS_ENDPOINT_LABEL = 'opendatahub.io/maas-endpoint';
+export const CONFIG_TYPE_LABEL = 'opendatahub.io/config-type';
+export const WELL_KNOWN_ANNOTATION = 'serving.kserve.io/well-known-config';
+export const DISABLED_ANNOTATION = 'opendatahub.io/disabled';
+export const ROUTING_TYPE_ANNOTATION = 'opendatahub.io/routing-type';
+export const SUPPORTED_TOPOLOGIES_ANNOTATION = 'opendatahub.io/supported-topologies';
+export const DASHBOARD_RESOURCE_LABEL = 'opendatahub.io/dashboard';
+export const VLLM_ADDITIONAL_ARGS = 'VLLM_ADDITIONAL_ARGS';
